requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
dnspython==2.8.0
fonttools==4.61.0
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.7
narwhals==2.13.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.5.0
pymongo==4.15.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
seaborn==0.13.2
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
